clotheshop-back/requirements.txt
Flask
Flask-Cors
psycopg2-binary
requests
